+++ crates/nostr-connect/src/lib.rs
@@ -15,3 +15,4 @@ pub mod client;
 pub mod error;
 pub mod prelude;
 pub mod signer;
+mod util;